+++ core/controller/site_content_tags.php
@@ -72,7 +72,7 @@ class site_content_tagsDocLister extends DocLister{
             $sysPlh=$this->renameKeyArr($this->_plh,$this->getCFGDef("sysKey","dl"));
 			$noneTPL=$this->getCFGDef("noneTPL","");
 			if(count($this->_docs)==0 && $noneTPL!=''){
-				$out=$this->modx->parseChunk($noneTPL,$sysPlh,"[+","+]");
+				$out=$this->parseChunk($noneTPL,$sysPlh);
 			}else{
                 if($this->checkExtender('user')){
                     $this->extender['user']->init($this,array('fields'=>$this->getCFGDef("userFields","")));
@@ -112,7 +112,7 @@ class site_content_tagsDocLister extends DocLister{
                     $class=implode(" ",$class);
                     $item[$this->getCFGDef("sysKey","dl").'.class']=$class;
 
-                    $tmp=$this->modx->parseChunk($tpl,$item,"[+","+]");
+                    $tmp=$this->parseChunk($tpl,$item);
 					if($this->getCFGDef('contentPlaceholder',0)!==0){
 						$this->toPlaceholders($tmp,1,"item[".$i."]"); // [+item[x]+] – individual placeholder for each iteration documents on this page
 					}
@@ -123,7 +123,7 @@ class site_content_tagsDocLister extends DocLister{
             $ownerTPL=$this->getCFGDef("ownerTPL","");
             // echo $this->modx->getChunk($ownerTPL);
             if($ownerTPL!=''){
-                $out=$this->modx->parseChunk($ownerTPL,array($this->getCFGDef("sysKey","dl").".wrap"=>$out),"[+","+]");
+                $out=$this->parseChunk($ownerTPL,array($this->getCFGDef("sysKey","dl").".wrap"=>$out));
             }
 		}else{
 			$out='none TPL';